curl fall back to wget
The implement of curl on Synology NAS is currently breaking up this script.

This PR adds to the config which download app (curl or wget) to use based on a file that is specific to Synology.

That said, the very first version I had made was adding a "curl check" and if it failed, then it would fall back to wget. Somehow I think this solution is still better since it can handle other odd curl on NAS or other Linux distribution.

Let me know what you think.